core: make thread ID a short int
Changes thread_get_id() and thread_get_id_may_fail() to return 'short
int' instead of 'int'. That is, 16 bits instead of 32 on all supported
architectures which is more than enough since the largest thread ID
value is (CFG_NUM_THREADS - 1). Note, struct wait_queue_elem::handle
is already a short int.

trace_ext_get_thread_id() is not changed (still returns an int) because
it is part of the TA API and modifying it would needlessly introduce
incompatibilities.

+++ core/arch/arm/include/kernel/thread.h
@@ -42,7 +42,7 @@ struct thread_core_local {
 	uint64_t x[4];
 #endif
 	vaddr_t tmp_stack_va_end;
-	int curr_thread;
+	short int curr_thread;
 	uint32_t flags;
 	vaddr_t abt_stack_va_end;
 #ifdef CFG_TEE_CORE_DEBUG
@@ -299,12 +299,12 @@ void thread_clr_boot_thread(void);
 /*
  * Returns current thread id.
  */
-int thread_get_id(void);
+short int thread_get_id(void);
 
 /*
  * Returns current thread id, return -1 on failure.
  */
-int thread_get_id_may_fail(void);
+short int thread_get_id_may_fail(void);
 
 /* Returns Thread Specific Data (TSD) pointer. */
 struct thread_specific_data *thread_get_tsd(void);
